requirements.txt
ffmpeg-python
whisper
yt_dlp
pytest
